requirements.txt
phidata
openai
sqlalchemy
duckduckgo-search
yfinance
fastapi
uvicorn
python-dotenv
qdrant-client
pypdf
ollama
asknews